feat(settings): add export-diagnostics action, remove dead support dashboard
The Support page is intentionally FAQ-only (ADR 0006). Delete the 22
unrendered diagnostics/health/recovery/knowledge-base components, their
13 unused support IPC contract methods + client/mock/hook entries, the
dead SUPPORT_TABS constants, and ~15 orphaned support.* i18n namespaces
from en + de.

Salvage the one capability with a real backend: add an "Export
diagnostics" action to Settings > About, wired through the
useExportDiagnostics service hook to the existing support_export_diagnostics
command. export_diagnostics is preserved end-to-end.

Resolves audit findings renderer-feat-3-001, p2-b1-ipc-chain-001,
p2-b1-ipc-chain-002, and a large share of the orphaned-i18n-key findings.
