--- apps/desktop/src/renderer/constants/support/index.ts
@@ -1,1 +1,0 @@
-export * from './support';
